Code Reorganisation and Adding RecordCount to Recommendations

+++ app/src/main/java/com/singhealth/enhance/activities/history/HistoryData.kt
@@ -11,5 +11,5 @@ data class HistoryData(
     val clinicDiaBPTarget: Long?,
     val clinicSysBP: Long?,
     val clinicDiaBP: Long?,
-    val scanRecordCount: Int?
+    val scanRecordCount: Long?
 )
